feat: Refactor monolithic app, fix bugs and build config
This commit completes a major refactoring of the React application, moving it from a single monolithic file to a clean, component-based architecture. It also resolves several critical bugs and configuration issues.

- Chore: Restructures project into a component-based architecture.
- Feat: Implements "remove friend" functionality.
- Fix: Corrects Tailwind CSS v4 configuration to resolve major styling issues. This includes updating dependencies, `vite.config.js`, and `src/index.css`, and removing `postcss.config.js`.
- Security: Removes hardcoded API key and configures it to be loaded from environment variables.
- Chore: Resolves all linting errors for a clean codebase.
