+++ src/context/ThemeContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const ThemeContext = createContext(null);
